Implement ConcatThenSplitIterDataset in PyGrain.
The algorithm concatenates elements in a buffer. When they don't fit anymore,
the remainder element is split, and the packed buffer is added to the output.
A notable optimization is that we can directly pass elements that already have
the desired sequence length to the output, as they are already packed.

The state is checkpointed everytime the buffer is empty or contains only one
remainder element. In order to restore any state, we restore to the last
checkpoint and then advance packing until the state is restored.

PiperOrigin-RevId: 729106464